+++ templates/index.html
@@ -1,0 +1,12 @@
+<head>
+  <title>Something</title>
+</head>
+<body>
+  <div>
+    <h1>Voici la transcription et traduction : </h1>
+    <p>ADN : {{ generationdata[0] }}</p>
+    <p>ARNm : {{ generationdata[1] }}</p>
+    <p>ARNm_START : {{ generationdata[2] }}</p>
+    <p>Protéine : {{ generationdata[3] }}</p>
+  </div>
+</body>
