+++ victor_suivi.doc
@@ -1,2 +1,3 @@
 - probleme index parcour de trad pour GEN
 - flask noninstall replit
+- change print into return forgot in if
